refactor: :recycle: move reset form to another file
